// Application/Frontend/frontend-app/src/generated/trade-client/model/watchlist-response.ts
/**
 * Trade REST API
 *
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */


export interface WatchlistResponse { 
    id: number;
    name: string;
    isDefault: boolean;
    symbols: Array<string>;
}